move: calculate_intrinsic_gas_after_merge
